refactor(socks): fold the policy route and resolver into the server
Four modules for one entry point was too many files to read: PolicyRoute and
TunnelResolver are helpers of Socks5Server and now live next to it. Header
comments that had grown into documentation are cut back to what the rest of
the tree carries.
